+++ packages/ingest/test/consumer.js
@@ -9,7 +9,8 @@ const { sleep } = require('@cumulus/common/util');
 const consumer = rewire('../consumer');
 const Consumer = consumer.Consumer;
 
-const timeToReceiveMessages = 1 * 1000;
+const timeToReceiveMessages = 200; // ms
+const timeLimitModifier = 50;
 let testConsumer;
 
 async function stubReceiveSQSMessages(_url, { numOfMessages }) {
@@ -17,36 +18,38 @@ async function stubReceiveSQSMessages(_url, { numOfMessages }) {
   return Array.apply(null, { length: numOfMessages }).map(() => 'i am a message'); // eslint-disable-line prefer-spread
 }
 consumer.__set__('receiveSQSMessages', stubReceiveSQSMessages);
-console.log(consumer.__get__('receiveSQSMessages'));
 consumer.__set__('deleteSQSMessage', async () => true);
 function processFn() {}
 
 const sandbox = sinon.sandbox.create();
-let processSpy;
+let batchSpy;
+let messageSpy;
 
 test.beforeEach(() => {
   // need to reinstantiate because this.now = Date.now()
   testConsumer = new Consumer();
-  processSpy = sandbox.spy(testConsumer, 'processMessage');
+  testConsumer.messageLimit = 40; // initial messagelimit
+  batchSpy = sandbox.spy(testConsumer, 'processMessages');
+  messageSpy = sandbox.spy(testConsumer, 'processMessage');
 });
 test.afterEach.always(() => sandbox.restore());
 
 test.serial('stops after timelimit', async (t) => {
-  testConsumer.messageLimit = 40;
-  testConsumer.timeLimit = timeToReceiveMessages * 2 - 100;
+  testConsumer.timeLimit = timeToReceiveMessages * 2 - timeLimitModifier;
 
   const result = await testConsumer.consume(processFn);
   t.is(result, 20);
-  t.is(processSpy.callCount, 20);
+  t.is(batchSpy.callCount, 2);
+  t.is(messageSpy.callCount, 20);
 });
 
 test.serial('continues when timeLimit is is greater than time to receive', async (t) => {
-  testConsumer.messageLimit = 40;
-  testConsumer.timeLimit = timeToReceiveMessages * 2 + 100;
+  testConsumer.timeLimit = timeToReceiveMessages * 2 + timeLimitModifier;
 
   const result = await testConsumer.consume(processFn);
   t.is(result, 30);
-  t.is(processSpy.callCount, 30);
+  t.is(batchSpy.callCount, 3);
+  t.is(messageSpy.callCount, 30);
 });
 
 test.serial('stops after messageLimit is reached', async (t) => {
@@ -55,7 +58,8 @@ test.serial('stops after messageLimit is reached', async (t) => {
 
   const result = await testConsumer.consume(processFn);
   t.is(result, 2);
-  t.is(processSpy.calledTwice, true);
+  t.is(batchSpy.calledOnce, true);
+  t.is(messageSpy.calledTwice, true);
 });
 
 test.serial('processMessages throws error on large batch sizes', async (t) => {
@@ -63,7 +67,7 @@ test.serial('processMessages throws error on large batch sizes', async (t) => {
     'Cannot process more than 10 messages per function call. Received limit: 20');
 });
 
-test.serial('processMessages respect messageLimit', async (t) => {
+test.serial('processMessages respects messageLimit', async (t) => {
   const result = await testConsumer.processMessages(processFn, 3);
   t.is(result, 3);
 });